+++ packages/types/src/model.ts
@@ -16,5 +16,8 @@ export interface ISpaceBase {
 }
 
 export interface ISpace extends ISpaceBase {
-  scheduledAt: string;
+  scheduledAt: null | string;
+}
+export interface ISpaceDocument extends ISpace {
+  id: string;
 }